refactor one class - one pyton file
