+++ tests/neuro_san/service/watcher/temp_networks/_test_base.py
@@ -22,7 +22,6 @@ default test-file pattern (test_*.py) skips it during collection.
 Concrete test classes live in sibling test_*.py files and inherit from
 S3ReservationsStorageTestBase.
 """
-import io
 import os
 import time
 
@@ -32,56 +31,11 @@ from typing import Dict
 from unittest import TestCase
 from unittest.mock import patch
 
-from botocore.exceptions import ClientError
-
 from neuro_san.internals.reservations.agent_reservation import AgentReservation
 from neuro_san.service.watcher.temp_networks.s3_reservations_storage \
     import S3ReservationsStorage
-
-
-# pylint: disable=invalid-name,unused-argument
-class FakeS3Client:
-    """
-    Minimal in-memory stand-in for a boto3 S3 client. Only implements the
-    methods that S3ReservationsStorage actually calls, storing object bodies
-    in a dict keyed by S3 object key. Method signatures use boto3's PascalCase
-    keyword arguments so the storage's call sites work unchanged.
-    """
-
-    def __init__(self):
-        """
-        Initialize an empty in-memory bucket.
-        """
-        self.objects: Dict[str, bytes] = {}
-
-    def head_bucket(self, Bucket: str):
-        """
-        Stand-in for boto3's head_bucket. Real boto3 returns a response dict
-        on success and raises ClientError otherwise. For tests we treat any
-        configured bucket as existing.
-        """
-        return {}
-
-    def put_object(self, Bucket: str, Key: str, Body, ContentType: str):
-        """
-        Store the given Body bytes (or str, encoded as utf-8) at Key.
-        """
-        if isinstance(Body, str):
-            Body = Body.encode("utf-8")
-        self.objects[Key] = Body
-        return {}
-
-    def get_object(self, Bucket: str, Key: str):
-        """
-        Return the stored bytes for Key wrapped in a Body stream, or raise
-        a NoSuchKey ClientError if the key is not present.
-        """
-        if Key not in self.objects:
-            raise ClientError(
-                {"Error": {"Code": "NoSuchKey", "Message": "Not Found"}},
-                "GetObject",
-            )
-        return {"Body": io.BytesIO(self.objects[Key])}
+from tests.neuro_san.service.watcher.temp_networks._fake_s3_client \
+    import FakeS3Client
 
 
 class S3ReservationsStorageTestBase(TestCase):
